second commit

+++ CMS_PROJECT/admin/functions.php
@@ -1,4 +1,52 @@
 <?php
+
+
+function escape($string) {
+    global $connection;
+    return mysqli_real_escape_string($connection, trim($string));
+}
+
+
+function users_online(){
+
+
+    if(isset($_GET['onlineusers'])){
+
+    
+
+    global $connection;
+
+    if(!$connection){
+        session_start();
+        require '../connection.php';
+        $session = session_id();
+        $time = time();
+        $time_out_in_seconds = 05;
+        $time_out = $time - $time_out_in_seconds;
+
+
+        $query = "SELECT * FROM users_online WHERE session = '$session'";
+        $send_query = mysqli_query($connection, $query);
+        $count = mysqli_num_rows($send_query);
+
+        if($count == NULL){
+            mysqli_query($connection, "INSERT INTO users_online(session, time) VALUES('$session','$time')");
+        }else{
+            mysqli_query($connection, "UPDATE users_online SET time = '$time' WHERE session = '$session'");
+        }
+        $users_online_query = mysqli_query($connection, "SELECT * FROM users_online WHERE time > '$time_out'");
+        echo $count_user = mysqli_num_rows($users_online_query);
+
+        }
+    
+ 
+        }   //Get Isset()
+}
+
+users_online();
+
+
+
 function confirm($result){
     global $connection;
     if(!$result){
@@ -16,19 +64,23 @@ function insert_categories()
     global $connection;
     if(isset($_POST['submit']))
     {
-        $cat_title = $_POST['cat_title'];
-        if($cat_title == "" || empty($cat_title))
-        {
-            echo "This feild should not be empty";
-        }
-        else
-        {
-            $query = "INSERT INTO categories(cat_title)";
-            $query .="VALUES('$cat_title')";
-            $create_category_query = mysqli_query($connection, $query);
-            if(!$create_category_query)
-            {
-                die('QUERY FAILED' . mysqli_error($connection));
+        if(isset($_SESSION['user_role'])){
+            if($_SESSION['user_role'] == 'admin'){
+                $cat_title = $_POST['cat_title'];
+                if($cat_title == "" || empty($cat_title))
+                {
+                echo "This feild should not be empty";
+                }
+                else
+                {
+                $query = "INSERT INTO categories(cat_title)";
+                $query .="VALUES('$cat_title')";
+                $create_category_query = mysqli_query($connection, $query);
+                if(!$create_category_query)
+                    {
+                        die('QUERY FAILED' . mysqli_error($connection));
+                    }
+                }
             }
         }
     }
@@ -64,14 +116,16 @@ function findAllCategories()
 function deleteCategories(){
     if(isset($_GET['delete']))
     {
-        global $connection;
-        $the_cate_id = $_GET['delete'];
-        $query = "DELETE FROM categories WHERE cat_id = $the_cate_id ";
-        $delete_query = mysqli_query($connection, $query);
-        header("location: categories.php");
-        exit();
-        
-
+        if(isset($_SESSION['user_role'])){
+            if($_SESSION['user_role'] == 'admin'){
+            global $connection;
+            $the_cate_id = $_GET['delete'];
+            $query = "DELETE FROM categories WHERE cat_id = $the_cate_id ";
+            $delete_query = mysqli_query($connection, $query);
+            header("location: categories.php");
+            exit();
+            }
+        }
     } 
 }
