+++ CMS_PROJECT/admin/view_all_users.php
@@ -63,7 +63,7 @@
             <td><?php echo "<a href='users.php?changed_to_admin={$user_id}'>Admin</a>" ?></td>
             <td><?php echo "<a href='users.php?changed_to_sub={$user_id}'>Subscriber</a>" ?></td>
             <td><?php echo "<a href='users.php?source=edit_user&edit_user={$user_id}'>Edit</a>" ?></td>
-            <td><?php echo "<a href='users.php?delete={$user_id}'>Delete</a>" ?></td>
+            <td><?php echo "<a onclick=\"javascript: return confirm('Are you sure to delete?'); \" href='users.php?delete={$user_id}'>Delete</a>" ?></td>
             
             
         </tr>
@@ -74,22 +74,30 @@
 
     <?php
         if(isset($_GET['changed_to_admin'])){
-            $the_user_id = $_GET['changed_to_admin'];
-            $query = "UPDATE users SET user_role= 'admin' WHERE user_id = $the_user_id ";
-            $change_to_admin_query = mysqli_query($connection, $query);
-            header("location: users.php");
-            exit();
+            if(isset($_SESSION['user_role'])){
+                if($_SESSION['user_role'] == 'admin'){
+                $the_user_id = $_GET['changed_to_admin'];
+                $query = "UPDATE users SET user_role= 'admin' WHERE user_id = $the_user_id ";
+                $change_to_admin_query = mysqli_query($connection, $query);
+                header("location: users.php");
+                exit();
+                }
+            }
         }
     ?>
 
 
     <?php
         if(isset($_GET['changed_to_sub'])){
-            $the_user_id = $_GET['changed_to_sub'];
-            $query = "UPDATE users SET user_role= 'subscribe' WHERE user_id = $the_user_id";
-            $change_to_sub_query = mysqli_query($connection, $query);
-            header("location: users.php");
-            exit();
+            if(isset($_SESSION['user_role'])){
+                if($_SESSION['user_role'] == 'admin'){
+                $the_user_id = $_GET['changed_to_sub'];
+                $query = "UPDATE users SET user_role= 'subscribe' WHERE user_id = $the_user_id";
+                $change_to_sub_query = mysqli_query($connection, $query);
+                header("location: users.php");
+                exit();
+                }
+            }
         }
     ?>
 
@@ -97,10 +105,15 @@
 
     <?php
         if(isset($_GET['delete'])){
-            $the_user_id = $_GET['delete'];
-            $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
-            $delete_user_query = mysqli_query($connection, $query);
-            header("location: users.php");
-            exit();
+            if(isset($_SESSION['user_role'])){
+                if($_SESSION['user_role'] == 'admin'){
+                    $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+                    $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+                    $delete_user_query = mysqli_query($connection, $query);
+                    header("location: users.php");
+                    
+                }
+            }
+          
         }
     ?>
